# performance(pie-modal): PEW-100 improves performance by deferring tasks