--- a/.changeset/shiny-coats-exercise.md
+++ b/.changeset/shiny-coats-exercise.md
@@ -1,0 +1,5 @@
+---
+"@justeattakeaway/pie-modal": minor
+---
+
+[Changed] - handleModalOpen method to help with performance