@xander-marjoram 2025-01-30
I'd be interested to see what the effect on the overall INP for the interaction is. Maybe if you record 5-10 before and after and compare the averages for each phase, that might emphasise the improvements.