@kevinrodrigues 2025-01-31
> I'd be interested to see what the effect on the overall INP for the interaction is. Maybe if you record 5-10 before and after and compare the averages for each phase, that might emphasise the improvements.

<img width="236" alt="Screenshot 2025-01-30 at 17 15 13" src="https://github.com/user-attachments/assets/b2b82857-0354-4e7f-8199-ca485f272069" />
<img width="241" alt="Screenshot 2025-01-30 at 17 36 21" src="https://github.com/user-attachments/assets/827bef00-d532-43cb-8aca-bb0543e17c8a" />